\name{DSDM3-package}
\alias{DSDM3-package}
\alias{DSDM3}
\docType{package}
\title{
\packageTitle{DSDM3}
}
\description{
\packageDescription{DSDM3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DSDM3}
\packageIndices{DSDM3}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{DSDM3}

Maintainer: \packageMaintainer{DSDM3}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
